+++ src/views/Zap/ZapBond.jsx
@@ -1,0 +1,128 @@
+import { useEffect, useState } from "react";
+import { useDispatch, useSelector } from "react-redux";
+import { formatCurrency, trim } from "../../helpers";
+import { Backdrop, Box, Fade, Grid, Paper, Tab, Tabs, Typography, Slide } from "@material-ui/core";
+import TabPanel from "../../components/TabPanel";
+import BondHeader from "../Bond/BondHeader";
+import BondRedeem from "../Bond/BondRedeem";
+import BondPurchase from "../Bond/BondPurchase";
+import "./zap.scss";
+import { useWeb3Context } from "src/hooks/web3Context";
+import { Skeleton } from "@material-ui/lab";
+
+function a11yProps(index) {
+  return {
+    id: `simple-tab-${index}`,
+    "aria-controls": `simple-tabpanel-${index}`,
+  };
+}
+
+function ZapBond({ bond }) {
+  const dispatch = useDispatch();
+  const { provider, address, chainID } = useWeb3Context();
+
+  const [slippage, setSlippage] = useState(0.5);
+  const [recipientAddress, setRecipientAddress] = useState(address);
+
+  const [view, setView] = useState(0);
+  const [quantity, setQuantity] = useState();
+
+  const isBondLoading = useSelector(state => state.bonding.loading ?? true);
+
+  const onRecipientAddressChange = e => {
+    return setRecipientAddress(e.target.value);
+  };
+
+  const onSlippageChange = e => {
+    return setSlippage(e.target.value);
+  };
+
+  useEffect(() => {
+    if (address) setRecipientAddress(address);
+  }, [provider, quantity, address]);
+
+  const changeView = (event, newView) => {
+    setView(newView);
+  };
+
+  return (
+    <div id="zap-view">
+      <Slide direction="left" in={true} mountOnEnter unmountOnExit>
+        <Paper className="ohm-card">
+          <BondHeader
+            bond={bond}
+            slippage={slippage}
+            recipientAddress={recipientAddress}
+            onSlippageChange={onSlippageChange}
+            onRecipientAddressChange={onRecipientAddressChange}
+            returnPath="zap/bond"
+          />
+
+          <Box direction="row" className="bond-price-data-row">
+            <div className="bond-price-data">
+              <Typography variant="h5" color="textSecondary">
+                Bond Price
+              </Typography>
+              <Typography variant="h3" className="price" color="primary">
+                {isBondLoading ? <Skeleton /> : formatCurrency(bond.bondPrice, 2)}
+              </Typography>
+            </div>
+            <div className="bond-price-data">
+              <Typography variant="h5" color="textSecondary">
+                Market Price
+              </Typography>
+              <Typography variant="h3" color="primary" className="price">
+                {isBondLoading ? <Skeleton /> : formatCurrency(bond.marketPrice, 2)}
+              </Typography>
+            </div>
+          </Box>
+
+          <Tabs
+            centered
+            value={view}
+            textColor="primary"
+            indicatorColor="primary"
+            onChange={changeView}
+            aria-label="bond tabs"
+          >
+            <Tab label="Bond" {...a11yProps(0)} />
+            <Tab label="Redeem" {...a11yProps(1)} />
+          </Tabs>
+
+          <TabPanel value={view} index={0}>
+            <BondPurchase bond={bond} slippage={slippage} recipientAddress={recipientAddress} />
+          </TabPanel>
+
+          <TabPanel value={view} index={1}>
+            <BondRedeem bond={bond} />
+          </TabPanel>
+        </Paper>
+      </Slide>
+    </div>
+  );
+}
+
+export function DisplayBondPrice({ bond }) {
+  const { chainID } = useWeb3Context();
+  return (
+    <>
+      {!bond.isAvailable[chainID] ? (
+        <>--</>
+      ) : (
+        `${new Intl.NumberFormat("en-US", {
+          style: "currency",
+          currency: "USD",
+          maximumFractionDigits: 2,
+          minimumFractionDigits: 2,
+        }).format(bond.bondPrice)}`
+      )}
+    </>
+  );
+}
+
+export function DisplayBondDiscount({ bond }) {
+  const { chainID } = useWeb3Context();
+  return <>{!bond.isAvailable[chainID] ? <>--</> : `${bond.bondDiscount && trim(bond.bondDiscount * 100, 2)}%`}</>;
+}
+
+export default ZapBond;
